small bug
git-svn-id: https://svn.generation.no/neatzsche/trunk@239 7a382ede-a22f-0410-891f-d2f9ed330ce3

+++ cpp/src/evaluator.cpp
@@ -32,7 +32,7 @@ double LightsimEvaluator::f(Phenotype * f)
       tmpin.push_back(j);
       vector <double> tmpout;
       tmpout=f->react(tmpin);
-      int tmpstrength=0;
+      double tmpstrength=0;
       int tmpwinner=0;
       for(unsigned int k=0;k<tmpout.size();k++){
 	if(tmpout.at(k)>tmpstrength&&tmpout.at(k)>0.3) {
